# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Cinema ticket booking system with payment integration",
        "title": "Cinema Booking API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/tikets/batch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tiket"
                ],
                "summary": "Create multiple tickets at once",
                "parameters": [
                    {
                        "description": "Batch tickets request",
                        "name": "batchRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateBatchTicketsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/tikets/{id}/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tiket"
                ],
                "summary": "Get detailed ticket information with film and jadwal details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tiket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TicketSummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CreateBatchTicketsRequest": {
            "type": "object",
            "properties": {
                "jadwal_id": {
                    "description": "The ID of the schedule for which tickets are being booked\nexample: 60d21b4667d0d8992e610c85",
                    "type": "string"
                },
                "kursis": {
                    "description": "Array of seat codes to book\nexample: [\"A1\", \"A2\", \"A3\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "description": "The ID of the user booking the tickets\nexample: 60d21b4667d0d8992e610c84",
                    "type": "string"
                }
            }
        },
...
        "models.TicketSummary": {
            "type": "object",
            "properties": {
                "film_poster": {
                    "description": "The URL to the film's poster image\nexample: https://example.com/posters/avengers.jpg",
                    "type": "string"
                },
                "film_title": {
                    "description": "The title of the film\nexample: Avengers: Endgame",
                    "type": "string"
                },
                "harga_tiket": {
                    "description": "The ticket price\nexample: 50000",
                    "type": "number"
                },
                "jadwal_ruangan": {
                    "description": "The theater room for the film\nexample: Studio 3",
                    "type": "string"
                },
                "jadwal_tanggal": {
                    "description": "The scheduled date for the film\nexample: 2025-07-25",
                    "type": "string"
                },
                "jadwal_waktu": {
                    "description": "The scheduled time for the film\nexample: 19:30",
                    "type": "string"
                },
                "status_pembayaran": {
                    "description": "The payment status\nexample: completed",
                    "type": "string"
                },
                "tanggal_pembayaran": {
                    "description": "The payment date (if paid)",
                    "type": "string"
                },
                "tiket": {
                    "description": "The ticket details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Tiket"
                        }
                    ]
                }
            }
        },
        "models.Tiket": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "The unique identifier of the ticket\nexample: 60d21b4667d0d8992e610c86",
                    "type": "string"
                },
                "created_at": {
                    "description": "The record creation timestamp",
                    "type": "string"
                },
                "jadwal_id": {
                    "description": "The ID of the schedule associated with this ticket\nexample: 60d21b4667d0d8992e610c85",
                    "type": "string"
                },
                "kursi": {
                    "description": "The seat code for this ticket\nexample: A1",
                    "type": "string"
                },
                "status": {
                    "description": "The ticket status (confirmed, cancelled, used)\nenum: confirmed,cancelled,used\nexample: confirmed",
                    "type": "string"
                },
                "tanggal_pembelian": {
                    "description": "The date and time when the ticket was purchased\nexample: 2025-07-22T15:04:05Z",
                    "type": "string"
                },
                "updated_at": {
                    "description": "The record last update timestamp",
                    "type": "string"
                },
                "user_id": {
                    "description": "The ID of the user who purchased the ticket\nexample: 60d21b4667d0d8992e610c84",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}